refactor(ex1): change parameter type, to throw exception.

+++ src/main/java/exercises/Exercise1.java
@@ -1,7 +1,7 @@
 package exercises;
 
 public class Exercise1 {
-    public static double divide(double numerator, double dominator) {
+    public static double divide(int numerator, int dominator) {
         return numerator / dominator;
     }
 }
